% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch arbuilder}
\usage{
launch(port = NULL, launch.browser = TRUE)
}
\arguments{
\item{port}{Port number. NULL for random.}

\item{launch.browser}{Open browser? Default TRUE.}
}
\description{
Launch arbuilder
}
